local/config: Remove use of CONFIG_HASH
`CONFIG_HASH` was loaded from the Keybase configuration file,
which is not actually guaranteed to exist on disk in a functional
Keybase instance (e.g., if Keybase was started using `keybase oneshot`).

To replace it, `STATUS_HASH` has been added. `STATUS_HASH` gets populated from
`keybase status -j`, and contains the active state of the Keybase instance.

This is a breaking change.